refactor(config): make config module private as labelme._config
Rename labelme/config to labelme/_config (and the unit test to
tests/unit/_config_test.py) to mark the config module as internal,
matching the existing private-module convention (_label_file, _shape,
_automation). This frees the config API from public-compatibility
constraints.

External usage on GitHub is limited to vendored source forks that copy
the entire tree (tests included); no project imports labelme.config as
a pip dependency, so this is not a real public API break.
