+++ README.md
@@ -113,7 +113,7 @@ labelme data_annotated/ --labels labels.txt  # specify label list with a file
 ### Command Line Arguments
 
 - `--output` specifies the location that annotations will be written to. If the location ends with .json, a single annotation will be written to this file. Only one image can be annotated if a location is specified with .json. If the location does not end with .json, the program will assume it is a directory. Annotations will be stored in this directory with a name that corresponds to the image that the annotation was made on.
-- The first time you run labelme, it will create a config file at `~/.labelmerc`. Add only the settings you want to override. For all available options and their defaults, see [`default_config.yaml`](labelme/config/default_config.yaml). If you would prefer to use a config file from another location, you can specify this file with the `--config` flag.
+- The first time you run labelme, it will create a config file at `~/.labelmerc`. Add only the settings you want to override. For all available options and their defaults, see [`default_config.yaml`](labelme/_config/default_config.yaml). If you would prefer to use a config file from another location, you can specify this file with the `--config` flag.
 - Without the `--nosortlabels` flag, the program will list labels in alphabetical order. When the program is run with this flag, it will display labels in the order that they are provided.
 - Flags are assigned to an entire image. [Example](examples/classification)
 - Labels are assigned to a single polygon. [Example](examples/bbox_detection)
@@ -145,7 +145,7 @@ pyinstaller labelme/labelme/__main__.py \
   --noconfirm \
   --specpath=build \
   --add-data=$(OSAM_PATH)/_models/yoloworld/clip/bpe_simple_vocab_16e6.txt.gz:osam/_models/yoloworld/clip \
-  --add-data=$(LABELME_PATH)/config/default_config.yaml:labelme/config \
+  --add-data=$(LABELME_PATH)/_config/default_config.yaml:labelme/_config \
   --add-data=$(LABELME_PATH)/icons/*:labelme/icons \
   --add-data=$(LABELME_PATH)/translate/*:translate \
   --icon=$(LABELME_PATH)/icons/icon-256.png \
